Refactored test folder
