+++ tests/checkout.test.jsx
@@ -1,9 +1,9 @@
 /* eslint-disable no-undef */
 import React from 'react';
-import { mount, shallow } from 'enzyme';
+import { shallow } from 'enzyme';
 import PropTypes from 'prop-types';
 import Checkout from '../client/components/checkout';
-import ShoeSizeSelect from '../client/components/checkoutSubComponents/shoeSizeSelect';
+import MockApp from './mockApp';
 
 describe('Rendering Tests', () => {
   const wrapper = shallow(<Checkout />);
@@ -14,44 +14,6 @@ describe('Rendering Tests', () => {
     expect(wrapper.find('buy-wrapper')).toBeDefined();
   });
 });
-
-const dummyProduct = {
-  productName: 'La Sportiva Tarantulace',
-  companyName: 'La Sportiva',
-  itemNumber: 830932,
-  color: 'Flame',
-  thumbnailUrl: 1,
-  price: '80.00',
-  rating: '4.3',
-  noRatings: 166,
-  shoeSizes: [42, 43, 44, 45],
-};
-
-class MockApp extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      images: [],
-      product: dummyProduct,
-      size: null,
-    };
-  }
-
-  render() {
-    const notMockFn = () => console.log('stupid science bitch cant make i more smarter');
-    const mock = notMockFn;
-    const handlers = {
-      shoeSizeSelect: mock,
-    };
-    const { product } = this.state;
-    return (
-      <Checkout
-        {...product}
-        handlers={handlers}
-      />
-    );
-  }
-}
 
 describe('When receiving props', () => {
   const wrapper = shallow(<MockApp />);
@@ -73,29 +35,6 @@ describe('When receiving props', () => {
   // Shoe Size Select
   it('should display the props', () => {
     expect(wrapper.contains(<h1>La Sportiv Tarantulace</h1>));
-  });
-});
-
-describe('Shoe Size Select', () => {
-  it('should receive shoe sizes', () => {
-    const wrap = shallow(<MockApp />);
-    expect(wrap.props()).toHaveProperty('shoeSizes');
-    expect(wrap.props('shoeSizes').shoeSizes).toHaveLength(4);
-  });
-
-  it('should render shoe size options', () => {
-    const wrap = shallow(<MockApp />);
-    expect(wrap.contains(<option value="42" key="42eu">42 Eu</option>));
-  });
-
-  it('should call a function onChange', () => {
-    const mockFn = jest.fn();
-    const wrap = shallow(<ShoeSizeSelect
-      handler={mockFn}
-      shoeSizes={dummyProduct.shoeSizes}
-    />);
-    wrap.find('select').simulate('change');
-    expect(mockFn).toHaveBeenCalled();
   });
 });
